auth_api/python/requirements.txt
click==8.0.1
Flask==2.0.1
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-client==0.0.1
mysql-connector==2.2.9
mysqlclient==2.0.3
PyJWT==2.1.0
Werkzeug==2.0.1